Provide AcmeVolumeDummy class for Sugar 0.98
The 0.98 desktop switched from GTK+ 2.x to 3.x, meaning that is this
mixer component used by the shell.  Unfortunately, we cannot use
AcmeVolumeGStreamer as we do with sugar-toolkit, as that patch must
use GStreamer 0.10 but 1.0 is used elsewhere, causing conflicting
type imports.

A better solution would be to provide an AcmeVolumePulseAudio backend,
but no such code exists (AcmeVolume last appeared in gnome-settings-daemon
2.24, over five years ago).  PTC.